+++ Beta Version/winsimple-controls.hpp
@@ -4,7 +4,7 @@
 
 #include <commctrl.h>   // for common controls (trackbar, etc.)
 #include <shlobj.h>     // for folder browser (BROWSEINFO, etc.)
-
+#include <filesystem>
 
 //Controls Linking: -lcomctl32 -lcomdlg32